testing triple_margin_loss over 5 epochs

+++ project/PatchDataset.py
@@ -25,7 +25,8 @@ class PatchDataset(Dataset):
         raw = raw.to(self.device)
 
         label = torch.from_numpy(io.imread('{}/{}'.format(self.label_path, patch_name)))
-        label = label.to(self.device, dtype=torch.int64)
+        # label = label.to(self.device, dtype=torch.int64)
+        label = label.to(self.device, dtype=torch.float)
 
         wmap = torch.from_numpy(io.imread('{}/{}'.format(self.wmap_path, patch_name)))
         wmap = wmap.to(self.device)
